spawn de inimigo 2,municao nas armas dos inimigos, e inimigo verifica a area (ele para no lugar e olha para as 4 direções)

+++ Assets/Scripts/ObjetosDoInventario/ArmaDeFogo.cs
@@ -33,7 +33,7 @@ public class ArmaDeFogo : ScriptableObject
     public bool RapidFire => rapidFire;
     public float RaioDoSomDoTiro => raioDoSomDoTiro;
     public float MunicaoCartucho => municaoCartucho;
-    public float Municao => municao;
+    public int Municao => municao;
     public Status GetStatus => GetStatusMetodo();
     public List<Melhoria> Melhorias => melhorias;
